Cleanup: Removing our faiss implementation (clustering + PCA) and using the original method instead
